requirements.txt
Django==5.1.2
djangorestframework==3.15.2
environs==11.0.0
packaging==24.1
pandas==2.2.3
drf-spectacular==0.27.2
pylint==3.3.1
graphene-django==3.2.2